FIX: Httpd::Module::Helpers::ModuleDetails Specs
Instead of extending the Httpd::Module::Helpers::ModuleDetails which added the DSL methods to the spec context object (whatever that is at the moment of the execution of the specs), I instead use the ModuleDetails object that was defined. This is important because all those `after_installing` method invokations were made on that object and that is what the `find_deletes` is searching through that.

I used a rspec let helper to define a subject and set that to the name of the module we care about testing for the information and then updated all the specs to call the `find_deletes` on that class.