resolving foodcritic FC005 and FC023
Resolving issue 79 which is to resolve FC005 and FC023 warnings.

Resolved all issues but 3 that are reported with regards to the provides block. These seem to be bogus as provides did not recognize the only_if or not_if chef DSL. Seems like FC005 check code needs to not validate or complain on provides block for custom resources.